+++ CHANGELOG.md
@@ -17,6 +17,7 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
     of sequences packed into a single bin.
   * Introduces `grain.experimental.RepeatIterDataset`.
   * Adds custom batching function support to `grain.DataLoader`.
+  * Adds `grain.experimental.FlatMapTransform` support to `grain.DataLoader`.
 
 * Breaking changes:
   * SliceMapDataset updated to use the full index relative to the parent